move slingshot to /startup
