+++ app/server/main.js
@@ -1,8 +1,8 @@
 import { Meteor } from 'meteor/meteor';
 
-import '/imports/lib/slingshot.js';
-
 import '/imports/startup/server/publications.js';
+import '/imports/startup/slingshot.js';
+
 import '/imports/api/methods.js';
 
 Meteor.startup(() => {
